fixed longpress on touch-devices
A timeout isnt NULL after it is cleared. the whole logic assumed this
would be the case. fix that.
Also prevent start of several Timeouts. When a Timeout is started while
another one is present we loose the handle to it and therefore cannot
clear it again. This lead to a lot of false-fired events.

fixes https://trello.com/c/e3Q9w2GM